[NS-72] Consolidate Spark SQL planning-related code
- Migrate all planning and plan transformation related code over to
`io.sparkcyclone.spark.transformation` package
