+++ src/main/scala/io/sparkcyclone/spark/plans/PlanCallsVeFunction.scala
@@ -1,6 +1,6 @@
 package io.sparkcyclone.spark.plans
 
-import io.sparkcyclone.spark.planning.VeFunction
+import io.sparkcyclone.spark.transformation.VeFunction
 import io.sparkcyclone.vectorengine.{LibraryReference, VeProcess}
 import org.apache.spark.sql.execution.SparkPlan
