+++ src/main/scala/io/sparkcyclone/spark/plans/VeAmplifyBatchesPlan.scala
@@ -3,7 +3,7 @@ package io.sparkcyclone.spark.plans
 import io.sparkcyclone.cache.ArrowEncodingSettings
 import io.sparkcyclone.data.vector.{VeBatchOfBatches, VeColBatch}
 import io.sparkcyclone.plugin.SparkCycloneExecutorPlugin.{source, veProcess, vectorEngine}
-import io.sparkcyclone.spark.planning.VeFunction
+import io.sparkcyclone.spark.transformation.VeFunction
 import io.sparkcyclone.util.CallContext
 import io.sparkcyclone.util.CallContextOps._
 import com.typesafe.scalalogging.LazyLogging
