+++ src/main/scala/io/sparkcyclone/spark/transformation/LibLocation.scala
@@ -1,10 +1,7 @@
-package io.sparkcyclone.spark.planning
-
-import java.nio.file.{Files, Path, Paths, StandardOpenOption}
+package io.sparkcyclone.spark.transformation
 
 import io.sparkcyclone.plugin.SparkCycloneExecutorPlugin
-import io.sparkcyclone.spark.{RequestCompiledLibraryForCode, RequestCompiledLibraryResponse}
-import org.apache.spark.api.plugin.PluginContext
+import java.nio.file.{Files, Path, Paths, StandardOpenOption}
 
 object LibLocation {
   trait LibLocation {
