+++ src/main/scala/io/sparkcyclone/spark/transformation/VeRewriteStrategyOptions.scala
@@ -1,4 +1,4 @@
-package io.sparkcyclone.spark.planning
+package io.sparkcyclone.spark.transformation
 
 import org.apache.spark.SparkConf
