+++ src/test/scala/io/sparkcyclone/tpc/TPCHVESqlSpec.scala
@@ -21,7 +21,7 @@ package io.sparkcyclone.tpc
 
 import io.sparkcyclone.plugin.{AuroraSqlPlugin, SparkCycloneExecutorPlugin}
 import io.sparkcyclone.plugin.SparkCycloneExecutorPlugin.CloseAutomatically
-import io.sparkcyclone.spark.planning.{VERewriteStrategy, VeRewriteStrategyOptions}
+import io.sparkcyclone.spark.transformation.{VERewriteStrategy, VeRewriteStrategyOptions}
 import io.sparkcyclone.vectorengine.VeProcess
 import java.io.File
 import org.apache.spark.sql.SparkSession
